Updated for recent Android SDK changes
$ANDROID_HOME is deprecated, use $ANDROID_SDK_ROOT instead.  But to
maintain compatibility with the older SDK, also set $ANDROID_HOME for
now.

The location of the SDK is different if Android Studio is installed with
v28 of the SDK, you pretty much have to use Android Studio if you want
ao GUI for creating/editing AVDs or for managing which SDK versions are
installed.  So, update the code to look for the SDK in the proper new
directory (at least on macOS, the Android Studio location might be
different for other systems, but I haven't checked yet).

Removed dependency of setting $GRADLE_HOME outside of if the Android SDK
is installed.  I suppose I might install it someday without the Android
SDK.

Finally, the v28 SDK moves the correct emulator binary to
${ANDROID_SDK_ROOT}/emulator, so add that to the path in front of
${ANDROID_SDK_ROOT}/tools, since the tools directory still has an
emulator binary (that doesn't work!).